+++ src/gl/gl.cpp
@@ -55,6 +55,8 @@ void gl_draw_movie_quad_common(uint32_t width, uint32_t height)
 	float movieWidth = ratio * width;
 	float movieOffsetY = (game_height - movieHeight) / 2.0f;
 
+	if (!ff8 && enable_time_cycle) newRenderer.setTimeFilterEnabled(true);
+
 	if (!ff8 && !ff7_field_center) movieOffsetY = 0.0f;
 
 	/*  y0    y2